exclude `test` directory while buiding
This will help us achieve lower package size on npm
