+++ package/src/components/test/ConnectButton.test.tsx
@@ -2,17 +2,6 @@ import React from 'react';
 import renderer, { act } from 'react-test-renderer';
 import { mainnet } from 'viem/chains';
 import { ConnectButton, KitProvider } from '../../index';
-import { requestObject } from './utils';
-
-declare global {
-  interface Window {
-    ethereum: {
-      on(arg0: string, arg1: (chainId: string) => void): unknown;
-      enable(): unknown;
-      request: typeof requestObject;
-    };
-  }
-}
 
 describe('test ConnectButton', () => {
   const Comp: React.FC = () => (
