+++ storybook-static/iframe.html
@@ -349,8 +349,8 @@
             window['DOCS_OPTIONS'] = {"defaultName":"Docs","autodocs":"tag"};</script><script type="module">import './sb-preview/runtime.js';
 
       
-      import './runtime~main.96457747.iframe.bundle.js';
+      import './runtime~main.a04039d0.iframe.bundle.js';
       
       import './807.670b5784.iframe.bundle.js';
       
-      import './main.91f232d7.iframe.bundle.js';</script></body></html>
+      import './main.0b8696a7.iframe.bundle.js';</script></body></html>
