+++ storybook-static/index.json
@@ -1,1 +1,1 @@
-{"v":4,"entries":{"pure-comps-connectbutton--primary":{"id":"pure-comps-connectbutton--primary","title":"Pure Comps/ConnectButton","name":"Primary","importPath":"./stories/BaseComponents/ConnectButton.stories.tsx","tags":["story"],"type":"story"},"pure-comps-modal--primary":{"id":"pure-comps-modal--primary","title":"Pure Comps/Modal","name":"Primary","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"type":"story"},"pure-comps-modal--close-on-overlay-click":{"id":"pure-comps-modal--close-on-overlay-click","title":"Pure Comps/Modal","name":"Close On Overlay Click","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"type":"story"},"pure-comps-switchnetworks--primary":{"id":"pure-comps-switchnetworks--primary","title":"Pure Comps/SwitchNetworks","name":"Primary","importPath":"./stories/BaseComponents/SwitchNetwork.stories.tsx","tags":["story"],"type":"story"},"compound-comps-connectbutton--primary":{"id":"compound-comps-connectbutton--primary","title":"Compound Comps/ConnectButton","name":"Primary","importPath":"./stories/CompoundComponents/ConnectButton.stories.tsx","tags":["story"],"type":"story"},"compound-comps-switchnetworks--primary":{"id":"compound-comps-switchnetworks--primary","title":"Compound Comps/SwitchNetworks","name":"Primary","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"type":"story"},"compound-comps-switchnetworks--custom-styling":{"id":"compound-comps-switchnetworks--custom-styling","title":"Compound Comps/SwitchNetworks","name":"Custom Styling","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"type":"story"}}}
+{"v":4,"entries":{"pure-comps-connectbutton--primary":{"id":"pure-comps-connectbutton--primary","title":"Pure Comps/ConnectButton","name":"Primary","importPath":"./stories/BaseComponents/ConnectButton.stories.tsx","tags":["story"],"type":"story"},"pure-comps-modal--primary":{"id":"pure-comps-modal--primary","title":"Pure Comps/Modal","name":"Primary","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"type":"story"},"pure-comps-modal--close-on-overlay-click":{"id":"pure-comps-modal--close-on-overlay-click","title":"Pure Comps/Modal","name":"Close On Overlay Click","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"type":"story"},"pure-comps-switchnetworks--primary":{"id":"pure-comps-switchnetworks--primary","title":"Pure Comps/SwitchNetworks","name":"Primary","importPath":"./stories/BaseComponents/SwitchNetwork.stories.tsx","tags":["story"],"type":"story"},"walletconnect-connectbutton--primary":{"id":"walletconnect-connectbutton--primary","title":"WalletConnect/ConnectButton","name":"Primary","importPath":"./stories/BaseComponents/WalletConnect.stories.tsx","tags":["story"],"type":"story"},"compound-comps-connectbutton--primary":{"id":"compound-comps-connectbutton--primary","title":"Compound Comps/ConnectButton","name":"Primary","importPath":"./stories/CompoundComponents/ConnectButton.stories.tsx","tags":["story"],"type":"story"},"compound-comps-switchnetworks--primary":{"id":"compound-comps-switchnetworks--primary","title":"Compound Comps/SwitchNetworks","name":"Primary","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"type":"story"},"compound-comps-switchnetworks--custom-styling":{"id":"compound-comps-switchnetworks--custom-styling","title":"Compound Comps/SwitchNetworks","name":"Custom Styling","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"type":"story"}}}
