+++ storybook-static/project.json
@@ -1,1 +1,1 @@
-{"generatedAt":1687288854068,"hasCustomBabel":false,"hasCustomWebpack":false,"hasStaticDirs":false,"hasStorybookEslint":true,"refCount":0,"packageManager":{"type":"pnpm","version":"8.6.0"},"framework":{"name":"@storybook/react-webpack5","options":{}},"builder":"@storybook/builder-webpack5","renderer":"@storybook/react","storybookVersion":"7.0.12","storybookVersionSpecifier":"^7.0.12","language":"typescript","storybookPackages":{"@storybook/blocks":{"version":"7.0.12"},"@storybook/react":{"version":"7.0.12"},"@storybook/react-webpack5":{"version":"7.0.12"},"@storybook/testing-library":{"version":"0.0.14-next.2"},"eslint-plugin-storybook":{"version":"0.6.12"},"storybook":{"version":"7.0.12"}},"addons":{"@storybook/addon-links":{"version":"7.0.12"},"@storybook/addon-essentials":{"version":"7.0.12"},"@storybook/addon-interactions":{"version":"7.0.12"}}}
+{"generatedAt":1687800112445,"hasCustomBabel":false,"hasCustomWebpack":false,"hasStaticDirs":false,"hasStorybookEslint":true,"refCount":0,"packageManager":{"type":"pnpm","version":"8.6.0"},"framework":{"name":"@storybook/react-webpack5","options":{}},"builder":"@storybook/builder-webpack5","renderer":"@storybook/react","storybookVersion":"7.0.12","storybookVersionSpecifier":"^7.0.12","language":"typescript","storybookPackages":{"@storybook/blocks":{"version":"7.0.12"},"@storybook/react":{"version":"7.0.12"},"@storybook/react-webpack5":{"version":"7.0.12"},"@storybook/testing-library":{"version":"0.0.14-next.2"},"eslint-plugin-storybook":{"version":"0.6.12"},"storybook":{"version":"7.0.12"}},"addons":{"@storybook/addon-links":{"version":"7.0.12"},"@storybook/addon-essentials":{"version":"7.0.12"},"@storybook/addon-interactions":{"version":"7.0.12"}}}
