+++ storybook-static/stories.json
@@ -1,1 +1,1 @@
-{"v":3,"stories":{"pure-comps-connectbutton--primary":{"id":"pure-comps-connectbutton--primary","title":"Pure Comps/ConnectButton","name":"Primary","importPath":"./stories/BaseComponents/ConnectButton.stories.tsx","tags":["story"],"kind":"Pure Comps/ConnectButton","story":"Primary","parameters":{"__id":"pure-comps-connectbutton--primary","docsOnly":false,"fileName":"./stories/BaseComponents/ConnectButton.stories.tsx"}},"pure-comps-modal--primary":{"id":"pure-comps-modal--primary","title":"Pure Comps/Modal","name":"Primary","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"kind":"Pure Comps/Modal","story":"Primary","parameters":{"__id":"pure-comps-modal--primary","docsOnly":false,"fileName":"./stories/BaseComponents/Modal.stories.tsx"}},"pure-comps-modal--close-on-overlay-click":{"id":"pure-comps-modal--close-on-overlay-click","title":"Pure Comps/Modal","name":"Close On Overlay Click","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"kind":"Pure Comps/Modal","story":"Close On Overlay Click","parameters":{"__id":"pure-comps-modal--close-on-overlay-click","docsOnly":false,"fileName":"./stories/BaseComponents/Modal.stories.tsx"}},"pure-comps-switchnetworks--primary":{"id":"pure-comps-switchnetworks--primary","title":"Pure Comps/SwitchNetworks","name":"Primary","importPath":"./stories/BaseComponents/SwitchNetwork.stories.tsx","tags":["story"],"kind":"Pure Comps/SwitchNetworks","story":"Primary","parameters":{"__id":"pure-comps-switchnetworks--primary","docsOnly":false,"fileName":"./stories/BaseComponents/SwitchNetwork.stories.tsx"}},"compound-comps-connectbutton--primary":{"id":"compound-comps-connectbutton--primary","title":"Compound Comps/ConnectButton","name":"Primary","importPath":"./stories/CompoundComponents/ConnectButton.stories.tsx","tags":["story"],"kind":"Compound Comps/ConnectButton","story":"Primary","parameters":{"__id":"compound-comps-connectbutton--primary","docsOnly":false,"fileName":"./stories/CompoundComponents/ConnectButton.stories.tsx"}},"compound-comps-switchnetworks--primary":{"id":"compound-comps-switchnetworks--primary","title":"Compound Comps/SwitchNetworks","name":"Primary","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"kind":"Compound Comps/SwitchNetworks","story":"Primary","parameters":{"__id":"compound-comps-switchnetworks--primary","docsOnly":false,"fileName":"./stories/CompoundComponents/SwitchNetwork.stories.tsx"}},"compound-comps-switchnetworks--custom-styling":{"id":"compound-comps-switchnetworks--custom-styling","title":"Compound Comps/SwitchNetworks","name":"Custom Styling","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"kind":"Compound Comps/SwitchNetworks","story":"Custom Styling","parameters":{"__id":"compound-comps-switchnetworks--custom-styling","docsOnly":false,"fileName":"./stories/CompoundComponents/SwitchNetwork.stories.tsx"}}}}
+{"v":3,"stories":{"pure-comps-connectbutton--primary":{"id":"pure-comps-connectbutton--primary","title":"Pure Comps/ConnectButton","name":"Primary","importPath":"./stories/BaseComponents/ConnectButton.stories.tsx","tags":["story"],"kind":"Pure Comps/ConnectButton","story":"Primary","parameters":{"__id":"pure-comps-connectbutton--primary","docsOnly":false,"fileName":"./stories/BaseComponents/ConnectButton.stories.tsx"}},"pure-comps-modal--primary":{"id":"pure-comps-modal--primary","title":"Pure Comps/Modal","name":"Primary","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"kind":"Pure Comps/Modal","story":"Primary","parameters":{"__id":"pure-comps-modal--primary","docsOnly":false,"fileName":"./stories/BaseComponents/Modal.stories.tsx"}},"pure-comps-modal--close-on-overlay-click":{"id":"pure-comps-modal--close-on-overlay-click","title":"Pure Comps/Modal","name":"Close On Overlay Click","importPath":"./stories/BaseComponents/Modal.stories.tsx","tags":["story"],"kind":"Pure Comps/Modal","story":"Close On Overlay Click","parameters":{"__id":"pure-comps-modal--close-on-overlay-click","docsOnly":false,"fileName":"./stories/BaseComponents/Modal.stories.tsx"}},"pure-comps-switchnetworks--primary":{"id":"pure-comps-switchnetworks--primary","title":"Pure Comps/SwitchNetworks","name":"Primary","importPath":"./stories/BaseComponents/SwitchNetwork.stories.tsx","tags":["story"],"kind":"Pure Comps/SwitchNetworks","story":"Primary","parameters":{"__id":"pure-comps-switchnetworks--primary","docsOnly":false,"fileName":"./stories/BaseComponents/SwitchNetwork.stories.tsx"}},"walletconnect-connectbutton--primary":{"id":"walletconnect-connectbutton--primary","title":"WalletConnect/ConnectButton","name":"Primary","importPath":"./stories/BaseComponents/WalletConnect.stories.tsx","tags":["story"],"kind":"WalletConnect/ConnectButton","story":"Primary","parameters":{"__id":"walletconnect-connectbutton--primary","docsOnly":false,"fileName":"./stories/BaseComponents/WalletConnect.stories.tsx"}},"compound-comps-connectbutton--primary":{"id":"compound-comps-connectbutton--primary","title":"Compound Comps/ConnectButton","name":"Primary","importPath":"./stories/CompoundComponents/ConnectButton.stories.tsx","tags":["story"],"kind":"Compound Comps/ConnectButton","story":"Primary","parameters":{"__id":"compound-comps-connectbutton--primary","docsOnly":false,"fileName":"./stories/CompoundComponents/ConnectButton.stories.tsx"}},"compound-comps-switchnetworks--primary":{"id":"compound-comps-switchnetworks--primary","title":"Compound Comps/SwitchNetworks","name":"Primary","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"kind":"Compound Comps/SwitchNetworks","story":"Primary","parameters":{"__id":"compound-comps-switchnetworks--primary","docsOnly":false,"fileName":"./stories/CompoundComponents/SwitchNetwork.stories.tsx"}},"compound-comps-switchnetworks--custom-styling":{"id":"compound-comps-switchnetworks--custom-styling","title":"Compound Comps/SwitchNetworks","name":"Custom Styling","importPath":"./stories/CompoundComponents/SwitchNetwork.stories.tsx","tags":["story"],"kind":"Compound Comps/SwitchNetworks","story":"Custom Styling","parameters":{"__id":"compound-comps-switchnetworks--custom-styling","docsOnly":false,"fileName":"./stories/CompoundComponents/SwitchNetwork.stories.tsx"}}}}
